+++ .env.example
@@ -1,2 +1,5 @@
 SECRET_KEY=secret_key
 DB_URL=postgresql://<username>:<password>@<hostname>:<port>/<db_name>
+
+MIGRATE_DB=false
+FLASK_APP=piss_tracker
