+++ .gitignore
@@ -1,1 +1,3 @@
 .env
+.python-version
+__pycache__/
