+++ Makefile
@@ -1,0 +1,5 @@
+start:
+	@docker-compose up -d
+
+stop:
+	@docker-compose down
