+++ README.md
@@ -1,0 +1,3 @@
+# Piss Tracker
+
+A Flask webapp meant to track your urination for people like me who do it way too much.
