--- app.py
@@ -1,44 +1,0 @@
-# https://www.digitalocean.com/community/tutorials/how-to-add-authentication-to-your-app-with-flask-login
-
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-from flask_login import LoginManager
-
-from dotenv import load_dotenv
-import os
-
-
-# init SQLAlchemy so we can use it later in our models
-db = SQLAlchemy()
-
-# def create_app():
-app = Flask(__name__)
-
-app.config['SECRET_KEY'] = os.getenv("SECRET_KEY")
-app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv("DB_URL")
-
-db.init_app(app)
-migrate = Migrate(app, db)
-
-login_manager = LoginManager()
-login_manager.login_view = 'auth.login'
-login_manager.init_app(app)
-
-from models import User
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(user_id)
-
-
-from auth import auth as auth_blueprint
-# blueprint for auth routes in our app
-app.register_blueprint(auth_blueprint)
-
-from main import main as main_blueprint
-# blueprint for non-auth parts of app
-app.register_blueprint(main_blueprint)
-
-
-if __name__ == "__main__":
-    app.run()
