--- main.py
@@ -1,14 +1,0 @@
-# https://www.digitalocean.com/community/tutorials/how-to-add-authentication-to-your-app-with-flask-login
-from flask import Blueprint, render_template
-from flask_login import login_required, current_user
-
-main = Blueprint('main', __name__)
-
-@main.route('/')
-def index():
-    return render_template('index.html')
-
-@main.route('/profile')
-@login_required
-def profile():
-    return render_template('profile.html', name=current_user.name)
