+++ src/piss_tracker/templates/profile.html
@@ -1,0 +1,8 @@
+{# https://www.digitalocean.com/community/tutorials/how-to-add-authentication-to-your-app-with-flask-login #}
+{% extends "base.html" %}
+
+{% block content %}
+<h1 class="title">
+  Welcome, {{ name }}!
+</h1>
+{% endblock %}
