--- templates/index.html
@@ -1,11 +1,0 @@
-{# https://www.digitalocean.com/community/tutorials/how-to-add-authentication-to-your-app-with-flask-login #}
-{% extends "base.html" %}
-
-{% block content %}
-<h1 class="title">
-  Flask Login Example
-</h1>
-<h2 class="subtitle">
-  Easy authentication and authorization in Flask.
-</h2>
-{% endblock %}
